==> Process/4-10 Dashboard Master.Rmd <==
---
title: 'Highmark CMS Stars Incentive Program 2018 Shiny Dashboard'
author: "Adam Sneath, Qing Wen (Maggie), and Pitt Analytics Team"
date: "March 25, 2019"
output: html_document
runtime: shiny
---

```{r message = FALSE}
library(shiny)
library(shinydashboard)
library(ggplot2)
library(ggmap)
library(leaflet)
library(leaflet.extras)
library(magrittr)
library(rgdal)
library(dplyr)
library(RColorBrewer)
library(DT)

DMD15 <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/Measures/DMD15.csv")
C12 <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/Measures/C12.csv")
C14 <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/Measures/C14.csv")
C15 <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/Measures/C15.csv")
C17 <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/Measures/C17.csv")
C20 <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/Measures/C20.csv")

DMD15Ranked <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/DMD15Ranked.csv")
C12Ranked <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/C12Ranked.csv")
C14Ranked <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/C14Ranked.csv")
C15Ranked <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/C15Ranked.csv")
C17Ranked <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/C17Ranked.csv")
C20Ranked <- read.csv("C:/Users/adsne/Desktop/2019 Spring/Healthcare/Project/C20Ranked.csv")
```

```{r Messing with Easy Button}
leaf <- leaflet() %>%
  addTiles() %>%
  addEasyButton(easyButton(
     icon = htmltools::span(class = "star", htmltools::HTML("&starf;")),
     onClick = JS("function(btn, map){ map.setZoom(1);}")))
leaf

leaf <- leaflet() %>%
  addTiles() %>%
  addEasyButtonBar(
   easyButton(
     icon = htmltools::span(class = "star", htmltools::HTML("&starf;")),
     onClick = JS("function(btn, map){ alert(\"Button 1\");}")),
   easyButton(
     icon = htmltools::span(class = "star", htmltools::HTML("&target;")),
     onClick = JS("function(btn, map){ alert(\"Button 2\");}")))
leaf

#OnClick = JS()

mapDMD15 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addEasyButton(easyButton(icon = htmltools::span(class = "star",htmltools::HTML("&starf;")),onClick = JS("function(btn, map){ map.setZoom(1);}"))) %>% addCircleMarkers(data = DMD15,lat=DMD15$Lat,lng=DMD15$Lon,weight=1, fillOpacity=1, radius = 1.9*log(DMD15$Denominator), color = ~palDMD15(DMD15$Compliance), popup = paste(DMD15$practice_Name,"<br>Denominator: ",as.character(DMD15$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",DMD15$Compliance), values = DMD15$Compliance, opacity = 0.7, title = "Compliance", position = "bottomright")

mapDMD15
```

```{r Messing with Scatter Plots}

scatterDMD15b <- ggplot(data=DMD15) + geom_point(aes(x=DMD15$Denominator,y=DMD15$Compliance))
scatterC12b <- ggplot(data=C12) + geom_point(aes(x=C12$Denominator,y=C12$Compliance))
scatterC14b <- ggplot(data=C14) + geom_point(aes(x=C14$Denominator,y=C14$Compliance))
scatterC15b <- ggplot(data=C15) + geom_point(aes(x=C15$Denominator,y=C15$Compliance))
scatterC17b <- ggplot(data=C17) + geom_point(aes(x=C17$Denominator,y=C17$Compliance))
scatterC20b <- ggplot(data=C20Ranked) + geom_point(aes(x=C20Ranked$Denominator,y=C20Ranked$Compliance))

scatterDMD15b
scatterC12b
scatterC14b
scatterC15b
scatterC17b
scatterC20b

```

```{r Runtime}
palDMD15 <- colorNumeric("RdYlGn",DMD15$Star_Rating)
palC12 <- colorNumeric("RdYlGn",C12$Star_Rating)
palC14 <- colorNumeric("RdYlGn",C14$Star_Rating)
palC15 <- colorNumeric("RdYlGn",C15$Star_Rating)
palC17 <- colorNumeric("RdYlGn",C17$Star_Rating)
palC20 <- colorNumeric("RdYlGn",C20$Star_Rating)

#Define plots here

mapDMD15 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addCircleMarkers(data=DMD15,lat=DMD15$Lat,lng=DMD15$Lon,weight=1, fillOpacity=1, radius = 1.9*log(DMD15$Denominator), color = ~palDMD15(DMD15$Star_Rating), popup = paste(DMD15$practice_Name,"<br>Denominator: ",as.character(DMD15$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",DMD15$Star_Rating), values = DMD15$Star_Rating, opacity = 0.7, title = "Star Rating", bins = 5, position = "bottomright")

mapC12 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addCircleMarkers(data = C12,lat=C12$Lat,lng=C12$Lon, weight=1, fillOpacity=1, radius = 2.3*log(1.75*C12$Denominator), color = ~palC12(C12$Star_Rating), popup = paste(C12$practice_Name,"<br>Denominator:",as.character(C12$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",C12$Star_Rating), values = C12$Star_Rating, opacity = 0.7, title = "Star Rating", bins = 5, position = "bottomright")

mapC14 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addCircleMarkers(data = C14, lat=C14$Lat,lng=C14$Lon, weight=1, fillOpacity=1, radius = 1.9*log(C14$Denominator), color = ~palC14(C14$Star_Rating), popup = paste(C14$practice_Name,"<br>Denominator: ",as.character(C14$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",C14$Star_Rating), values = C14$Star_Rating, opacity = 0.7, title = "Star Rating", bins = 5, position = "bottomright")

mapC15 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addCircleMarkers(data = C15, lat=C15$Lat,lng=C15$Lon, weight=1, fillOpacity=1, radius = 1.9*log(C15$Denominator), color = ~palC15(C15$Star_Rating), popup = paste(C15$practice_Name,"<br>Denominator: ",as.character(C15$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",C15$Star_Rating), values = C15$Star_Rating, opacity = 0.7, title = "Star Rating", bins = 5, position = "bottomright")

mapC17 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addCircleMarkers(data = C17, lat=C17$Lat,lng=C17$Lon, weight=1, fillOpacity=1, radius = 2.1*log(1.75*C17$Denominator), color = ~palC17(C17$Star_Rating), popup = paste(C17$practice_Name,"<br>Denominator: ",as.character(C17$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",C17$Star_Rating), values = C17$Star_Rating, opacity = 0.7, title = "Star Rating", bins = 5, position = "bottomright")

mapC20 <- leaflet() %>% addTiles() %>% addSearchOSM() %>% addCircleMarkers(data = C20, lat=C20$Lat,lng=C20$Lon, weight=1, fillOpacity=1, radius = 1.9*log(C20$Denominator), color = ~palC20(C20$Star_Rating), popup = paste(C20$practice_Name,"<br>Denominator: ",as.character(C20$Denominator))) %>% addLegend(pal = colorNumeric("RdYlGn",C20$Star_Rating), values = C20$Star_Rating, opacity = 0.7, title = "Star Rating", bins = 5, position = "bottomright")

scatterDMD15 <- ggplot(data=DMD15) + geom_point(aes(x=DMD15$Denominator,y=DMD15$Compliance))

scatterC12 <- ggplot(data=C12) + geom_point(aes(x=C12$Denominator,y=C12$Compliance))

scatterC14 <- ggplot(data=C14) + geom_point(aes(x=C14$Denominator,y=C14$Compliance))

scatterC15 <- ggplot(data=C15) + geom_point(aes(x=C15$Denominator,y=C15$Compliance))

scatterC17 <- ggplot(data=C17) + geom_point(aes(x=C17$Denominator,y=C17$Compliance))

scatterC20 <- ggplot(data=C20) + geom_point(aes(x=C20$Denominator,y=C20$Compliance))

histoDMD15 <- ggplot(DMD15Ranked, aes(x=Compliance)) + geom_histogram(bins = 10,col="black",fill="deepskyblue3") + ggtitle("Where Do You Stand?")

histoC12 <- ggplot(C12Ranked, aes(x=Compliance)) + geom_histogram(bins = 10,col="black",fill="deepskyblue3") + ggtitle("Where Do You Stand?")

histoC14 <- ggplot(C14Ranked, aes(x=Compliance)) + geom_histogram(bins = 10,col="black",fill="deepskyblue3") + ggtitle("Where Do You Stand?")

histoC15 <- ggplot(C15Ranked, aes(x=Compliance)) + geom_histogram(bins = 10,col="black",fill="deepskyblue3") + ggtitle("Where Do You Stand?")

histoC17 <- ggplot(C17Ranked, aes(x=Compliance)) + geom_histogram(bins = 10,col="black",fill="deepskyblue3") + ggtitle("Where Do You Stand?")

histoC20 <- ggplot(C20Ranked, aes(x=Compliance)) + geom_histogram(bins = 10,col="black",fill="deepskyblue3") + ggtitle("Where Do You Stand?")

tabDMD15 <- datatable(DMD15Ranked,extensions = "Responsive",filter="top", caption="Showing Denominator > 2")

tabC12 <- datatable(C12Ranked,extensions = "Responsive",filter="top", caption="Showing Denominator > 2")

tabC14 <- datatable(C14Ranked,extensions = "Responsive",filter="top", caption="Showing Denominator > 2")

tabC15 <- datatable(C15Ranked,extensions = "Responsive",filter="top", caption="Showing Denominator > 2")

tabC17 <- datatable(C17Ranked,extensions = "Responsive",filter="top", caption="Showing Denominator > 2")

tabC20 <- datatable(C20Ranked,extensions = "Responsive",filter="top", caption="Showing Denominator > 2")

bul <- "\U2022"

bestDMD15 <- renderUI({tags$div(tags$h4(paste(bul,"Initiate conversation with the patient")),tags$h4(paste(bul,"Review purpose of the medication, and emphasize its importance")),tags$h4("Strategies to improve adherence:"),tags$h4("1.) recommend a 90-day supply"),tags$h4("2.) encourage the patient to enroll in an auto-fill program"), tags$h4("3.) suggest reminder tools such as alarms, a specific storage plan for medication"))})

bestC12 <- renderUI({tags$div(tags$h4(paste(bul,"Taking proper medication")),tags$h4(paste(bul,"Bisphosphonate Therapy has been proved to be very effective. Bisphosphonates are used to treat thin/fragile bones and helps prevent future bone fractures.")), tags$h4(paste(bul,"Teriparatide and denosumab are effective alternative medication for bisphosphonates.")),tags$h4(paste(bul,"Dual energy x-ray absorptiometry screenings for all women 65+ years old is recommended by US Preventative Services Task Force")),tags$h4(paste(bul,"FRAX Fracture Risk Assessment Tool determines fracture risk for younger women")))})

bestC14 <- renderUI({tags$div(tags$h4(paste(bul,"Possible prescription of RAAS-inhibiting drugs")),tags$h4(paste(bul,"Glycemic Control")),tags$h4(paste(bul,"Blood Pressure Management")),tags$h4(paste(bul,"Cholestorol level Management")),tags$h4(paste(bul,"Dietary Sodium Restriction")),tags$h4(paste(bul,"Diuretic Therapy")))})

bestC15 <- renderUI({tags$div(tags$h4(paste(bul,"Set appropriate glycemic targets.")),tags$h4(paste(bul,"Note the progressiveness of Type 2 diabetes - what works now may not work tomorrow. ")),tags$h4(paste(bul,"Culturally specific care and education (diabetes more prevalent in minority populations")),tags$h4(paste(bul,"Provide access to all necessary services at one visit to elminate barriers of travel and transportation")))})

bestC17 <- renderUI({tags$div(tags$h4(paste(bul,"Tight control can prevent or reduce the rate of joint damage.")),tags$h4(paste(bul,"Better educate patients on importance of taking proper medication to motivate them to take their prescriptions regularly.")),tags$h4(paste(bul,"Methotrexate: most commonly used DMARD; shown to work better than any other single medicine.")),tags$h4(paste(bul, "Minocycline (Minocin): antibiotic that helps reduce and cease inflammation")))}) 

bestC20 <- renderUI({tags$div(tags$h4(paste(bul,"Prompt patient to bring list of current medication")),tags$h4(paste(bul,"Text/Email remind them to bring medication list")),tags$h4(paste(bul,"Have a computer system in place that keeps track of patients current medications")),tags$h4(paste(bul,"Text/Email remind patients to take medication and maintain healthy lifestyle")))})

defDMD15 <- renderUI({tags$h4("The percentage of Members age 40-75 who were dispensed a medication for diabetes that receive a statin medication.")})
  
defC12 <- renderUI({tags$h4("The percentage of female Members age 67 - 85 who suffered a fracture and who had either a bone mineral density (BMD) test or prescription for a drug to treat or prevent osteoporosis in the six months after the fracture.")})

defC14 <- renderUI({tags$h4("The percentage of diabetic Members age 18 - 75 who had medical attention for nephropathy.")})

defC15 <- renderUI({tags$h4("The percentage of Members age 18 - 75 with diabetes (type 1 or type 2) whose most recent Hemoglobin A1C is <=9.0%.")})

defC17 <- renderUI({tags$h4("The percentage of Members age 18 and older who were diagnosed with rheumatoid arthritis and who were dispensed at least one ambulatory prescription for a disease-modifying anti-rheumatic drug (DMARD).")})

defC20 <- renderUI({tags$h4("The percentage of discharges from January 1-December 1 of the measurement year for Members 18 years of age and older for whom medications were reconciled the date of discharge through 30 days after discharge (31 total days).")})

tips <- renderUI({tags$div(tags$h4("~ Use the Dashboard for analysis or visual aid during presentation."),tags$h4("~ Dashboard is interactive, try zooming & searching on the map, or searching the data table."),tags$h4("~ Can zoom entire Dashboard UI with pinch on laptop touch pad - want to add this feature for mouse users."),tags$h4("~ Map point size is best at mid-level zoom: zoom in a few clicks from default, or out a few clicks after searching for a specific location (which auto-zooms to street-level)."),tags$h4("~ Click map points to see practice name & practice denominator in small popup window. The name can be copy-pasted into the data table for more analysis."),tags$h4("~ Data table can be sorted (ascending or descending) for any column."),tags$h4("~ Most useful data table searches are city, or practice name."),tags$h4("~ The 'City.Rank' column is computed with a simple algorithm that ranks each practice within it's 'City' as defined in the data."),tags$h4("~ Regression analysis is ongoing; so far it seems that small denominator practices tend to have lower compliance, which may simply be feature of sample size, rather than actual indicator of poor performance."),tags$h4("~ The table and histograms only display practices with denominator > 2 to avoid excess 0s and 1s (for compliance) from tiny denominator practices. This adjustment intends to more clearly demonstrates the compliance distributions."))})

### Load Dashboard Starts Here ### Can make this a separate code block for ease of editing definitions vs. runtime

ui <- dashboardPage(
  dashboardHeader(
    title = "2018 CMS Stars Dashboard"
  ),
  dashboardSidebar(  
    sidebarMenu(
      menuItem("Dashboard Introduction", tabName = "introduction", icon = icon("dashboard")),
      menuItem("DMD 15", tabName = "dmd15", icon = icon("stethoscope"), badgeLabel = "Bad", badgeColor = "red"),
      menuItem("C 12", tabName = "C12", icon = icon("stethoscope"), badgeLabel = "Bad", badgeColor = "red"),
      menuItem("C 14", tabName = "c14", icon = icon("stethoscope"), badgeLabel = "Bad", badgeColor = "red"),
      menuItem("C 15", tabName = "c15", icon = icon("stethoscope"), badgeLabel = "Bad", badgeColor = "red"),
      menuItem("C 17", tabName = "c17", icon = icon("stethoscope"), badgeLabel = "Bad", badgeColor = "red"),
      menuItem("C 20", tabName = "C20", icon = icon("stethoscope"), badgeLabel = "Bad", badgeColor = "red")
      )
  ),
  dashboardBody(
    tabItems(
      tabItem(tabName = "introduction", h1("About"),
              fluidRow(box(title = "Helpful Tips", status = "primary", solidHeader = T,background = "aqua",uiOutput("tips"),width=12)
              )
      ),
      tabItem(tabName = "dmd15", h1("DMD15 Statin Use in Persons with Diabetes"),
              fluidRow(box(title = "Definition", status = "primary", solidHeader = T,background = "aqua",uiOutput("defDMD15"),width=12)
              ),
              fluidRow(box(title = "Compliance by Location", status = "primary", solidHeader = T, background = "aqua",leafletOutput("mapDMD15")),
                       box(title = "Compliance Histogram", status = "primary", solidHeader = T,background = "aqua",plotOutput("histoDMD15"))
              ),
              fluidRow(box(DT::dataTableOutput('tabDMD15'),width = 12)
              ),
              fluidRow(box(title = "Best Practices", status = "primary", solidHeader = T,background = "aqua",uiOutput("bestDMD15")),
                       box(title = "Denominator as Predictor", status = "primary", solidHeader = T,background = "aqua",plotOutput("regDMD15")) 
              )
      ),
      tabItem(tabName = "C12", h1(" C12 Osteoporosis Management in Women who had a Fracture"),
              fluidRow(box(title = "Definition", status = "primary", solidHeader = T,background = "aqua",uiOutput("defC12"),width=12)
              ),
              fluidRow(box(title = "Compliance by Location", status = "primary", solidHeader = T, background = "aqua",leafletOutput("mapC12")),
                       box(title = "Compliance Histogram", status = "primary", solidHeader = T,background = "aqua",plotOutput("histoC12"))
              ),
              fluidRow(box(DT::dataTableOutput('tabC12'),width = 12)
              ),
              fluidRow(box(title = "Best Practices", status = "primary", solidHeader = T,background = "aqua",uiOutput("bestC12")),
                       box(title = "Denominator as Predictor", status = "primary", solidHeader = T,background = "aqua",plotOutput("regC12")) 
              )
      ),
      tabItem(tabName = "c14", h1("C14 Comprehensive Diabetes Care: Medical Attention for Nephropathy"),
              fluidRow(box(title = "Definition", status = "primary", solidHeader = T,background = "aqua",uiOutput("defC14"),width=12)
              ),
              fluidRow(box(title = "Compliance by Location", status = "primary", solidHeader = T, background = "aqua",leafletOutput("mapC14")),
                       box(title = "Compliance Histogram", status = "primary", solidHeader = T,background = "aqua",plotOutput("histoC14"))
              ),
              fluidRow(box(DT::dataTableOutput('tabC14'),width = 12)
              ),
              fluidRow(box(title = "Best Practices", status = "primary", solidHeader = T,background = "aqua",uiOutput("bestC14")),
                       box(title = "Denominator as Predictor", status = "primary", solidHeader = T,background = "aqua",plotOutput("regC14")) 
              )
      ),
      tabItem(tabName = "c15", h1("C15 Comprehensive Diabetes Care: HbA1c Control (<=9%) (DYNAMIC)"),
              fluidRow(box(title = "Definition", status = "primary", solidHeader = T,background = "aqua",uiOutput("defC15"),width=12)
              ),
              fluidRow(box(title = "Compliance by Location", status = "primary", solidHeader = T, background = "aqua",leafletOutput("mapC15")),
                       box(title = "Compliance Histogram", status = "primary", solidHeader = T,background = "aqua",plotOutput("histoC15"))
              ),
              fluidRow(box(DT::dataTableOutput('tabC15'),width = 12)
              ),
              fluidRow(box(title = "Best Practices", status = "primary", solidHeader = T,background = "aqua",uiOutput("bestC15")),
                       box(title = "Denominator as Predictor", status = "primary", solidHeader = T,background = "aqua",plotOutput("regC15")) 
              )
      ),
      tabItem(tabName = "c17", h1("C17 Disease-Modifying Anti-Rheumatic Drug Therapy for Rheumatoid Arthritis"),
              fluidRow(box(title = "Definition", status = "primary", solidHeader = T,background = "aqua",uiOutput("defC17"),width=12)
              ),
              fluidRow(box(title = "Compliance by Location", status = "primary", solidHeader = T, background = "aqua",leafletOutput("mapC17")),
                       box(title = "Compliance Histogram", status = "primary", solidHeader = T,background = "aqua",plotOutput("histoC17"))
              ),
              fluidRow(box(DT::dataTableOutput('tabC17'),width = 12)
              ),
              fluidRow(box(title = "Best Practices", status = "primary", solidHeader = T,background = "aqua",uiOutput("bestC17")),
                       box(title = "Denominator as Predictor", status = "primary", solidHeader = T,background = "aqua",plotOutput("regC17")) 
              )
      ),
      tabItem(tabName = "C20", h1("C20 Medication Reconciliation Post-Discharge"),
              fluidRow(box(title = "Definition", status = "primary", solidHeader = T,background = "aqua",uiOutput("defC20"),width=12)
              ),
              fluidRow(box(title = "Compliance by Location", status = "primary", solidHeader = T, background = "aqua",leafletOutput("mapC20")),
                       box(title = "Compliance Histogram", status = "primary", solidHeader = T,background = "aqua",plotOutput("histoC20"))
              ),
              fluidRow(box(DT::dataTableOutput('tabC20'),width = 12)
              ),
              fluidRow(box(title = "Best Practices", status = "primary", solidHeader = T,background = "aqua",uiOutput("bestC20")),
                       box(title = "Denominator as Predictor", status = "primary", solidHeader = T,background = "aqua",plotOutput("regC20")) 
              )
      )
    )
  )
)

server <- function(input,output) {
  #Instructions
  output$tips <- tips
  #Definitions
  output$defDMD15 <- defDMD15
  output$defC12 <- defC12
  output$defC14 <- defC14
  output$defC15 <- defC15
  output$defC17 <- defC17
  output$defC20 <- defC20
  #Data Tables
  output$tabDMD15 <- DT::renderDataTable({tabDMD15})
  output$tabC12 <- DT::renderDataTable({tabC12})
  output$tabC14 <- DT::renderDataTable({tabC14})
  output$tabC15 <- DT::renderDataTable({tabC15})
  output$tabC17 <- DT::renderDataTable({tabC17})
  output$tabC20 <- DT::renderDataTable({tabC20})
  #Map
  output$mapDMD15 <- renderLeaflet({mapDMD15})
  output$mapC12 <- renderLeaflet({mapC12})
  output$mapC14 <- renderLeaflet({mapC14})
  output$mapC15 <- renderLeaflet({mapC15})
  output$mapC17 <- renderLeaflet({mapC17})
  output$mapC20 <- renderLeaflet({mapC20})
  #Compliance Histo
  output$histoDMD15 <- renderPlot({histoDMD15})
  output$histoC12 <- renderPlot({histoC12})
  output$histoC14 <- renderPlot({histoC14})
  output$histoC15 <- renderPlot({histoC15})
  output$histoC17 <- renderPlot({histoC17})
  output$histoC20 <- renderPlot({histoC20})
  #Scatter Plot
  output$regDMD15 <- renderPlot({scatterDMD15})
  output$regC12 <- renderPlot({scatterC12})
  output$regC14 <- renderPlot({scatterC14})
  output$regC15 <- renderPlot({scatterC15})
  output$regC17 <- renderPlot({scatterC17})
  output$regC20 <- renderPlot({scatterC20})
  #Best Practices 
  output$bestDMD15 <- bestDMD15
  output$bestC12 <- bestC12
  output$bestC14 <- bestC14
  output$bestC15 <- bestC15
  output$bestC17 <- bestC17
  output$bestC20 <- bestC20
  
}

shinyApp(ui,server)

```
